refactor: client side에있던 서버타입 server.types로 이동
