+++ types/server.types.ts
@@ -28,3 +28,34 @@ export type MultiReviewData = {
 export type SpotData = {
   averageStars: Stars;
 } & Destination;
+
+export type GuestUpdateType = {
+  birthDate: string;
+  gender: string;
+  nickName: string;
+};
+
+export type GetGuestUpdateType = {
+  nickName: string;
+  imageUrl: string;
+  gender: string;
+  birthDate: string;
+  role: string;
+};
+
+export type LoginOauthType = {
+  id: number;
+  nickName: string;
+  imageUrl: string;
+  role: string; //ROLE_USER, ROLE_GUEST
+  tokenType: string;
+  accessToken: string;
+};
+
+export type GetUserInfoType = {
+  userId: number;
+  nickName: string;
+  gender: string;
+  age: number;
+  imageUrl: string;
+};
